refactor: move webhook stuff to new class
